fix(Byte): update isHex validator (#543)
* fix(Byte): replace validators with explicit type

The base64 and hex validators are not guaranteeed to work.
There are some false positive cases, so instead add a type header to
the string, so a user can specify what the encoding of the string is.

* fix(Byte): update hex validator

There were cases where the hex regex validator would return true
for a value that was acutally base64 encoded. Added tests for this
case and updated the validator.