feat(scalars): parse objectid using bson
The `ObjectID` scalar type does some great work ensuring that the data
in the database maps to valid ObjectID values for MongoDB. However, in
order to use arguments supplied as ObjectID, we still need to convert
them to a `bson.ObjectId` type on-the-fly in resolvers or else we'll end
up not matching documents that should be matched.

This change adds a dependency on the `bson` library so we can use its
`ObjectId` class as a means of serializing ObjectID data from the
GraphQL API into a usable object for resolvers.

Example of code before this change:

```javascript
someResolver: (_parent, { id }, { db }) => {
  const someData = await db.collection('things').findOne({ _id: new ObjectId(id) })

  return someData
},
```

And here's what it will look like afterward:

```javascript
someResolver: (_parent, { id }, { db }) => {
  const someData = await db.collection('things').findOne({ _id: id })

  return someData
},
```

Similar to `Date` objects which are serialized appopriately in MongoDB
if you use `Timestamp`, ObjectIDs should be parsed properly into their
correct type before persistence into the database. By doing so, we're
ensuring that this type is always consistent.

Resolves #429