fix: DateTime: timestamps are expected to be number of milliseconds
graphql-iso-date operated on Unix timestamp values, but graphql-scalars operates on ECMAScript timestamps (number of milliseconds since January 1, 1970, UTC)
as decided in https://github.com/Urigo/graphql-scalars/issues/387#issuecomment-648202200

It has to be clear which is used. Certainly values are not Unix timestamps and all references must be removed.
Docs are updated.

ref: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#the_ecmascript_epoch_and_timestamps